+++ packages/extras/src/lib/components/MeshDiscardMaterial/MeshDiscardMaterial.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import type { MeshDiscardMaterialProps } from './types'
-  import { fragmentShader } from './fragment'
-  import { vertexShader } from './vertex'
+  import type { MeshDiscardMaterialProps } from './types.js'
+  import { fragmentShader } from './fragment.js'
+  import { vertexShader } from './vertex.js'
   import { T } from '@threlte/core'
 
   let { children, ref = $bindable(), ...props }: MeshDiscardMaterialProps = $props()