+++ packages/rapier/src/lib/components/Attractor/Attractor.svelte
@@ -2,8 +2,8 @@
   import type { RigidBody } from '@dimforge/rapier3d-compat'
   import { T, useTask } from '@threlte/core'
   import { Group, Vector3 } from 'three'
-  import { useRapier } from '../../hooks/useRapier'
-  import type { AttractorProps } from './types'
+  import { useRapier } from '../../hooks/useRapier.js'
+  import type { AttractorProps } from './types.js'
 
   let {
     strength = 1,