+++ packages/xr/src/lib/components/VRButton.svelte
@@ -12,7 +12,7 @@
 <script lang="ts">
   import type { ComponentProps } from 'svelte'
   import XRButton from './XRButton.svelte'
-  import { defaultFeatures } from '../internal/defaultFeatures'
+  import { defaultFeatures } from '../internal/defaultFeatures.js'
 
   type Props = Omit<ComponentProps<typeof XRButton>, 'mode' | 'sessionInit'>