+++ packages/studio/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @threlte/studio
 
+## 0.2.0
+
+### Minor Changes
+
+- ffc74ea: update "module" and "moduleResolution" to "NodeNext" and adjust all relative imports
+
 ## 0.1.11
 
 ### Patch Changes